+++ client/CHANGELOG.md
@@ -25,12 +25,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `GetProtocolParameters` to the message interface;
 - Max length checks for transaction payload and essence, so the block size isn't exceeded, even with 8 parents;
+- `{PreparedTransactionData, SignedTransactionData, RemainderData, InputSigningData}::try_from_dto_unverified`;
 
 ### Changed
 
 - Update dependencies;
 - Use new `iota-types` crate instead of `bee-api-types` and `bee-block`;
 - Use new `iota-pow` crate instead of `bee-pow`;
+- `Message::{SignTransaction, BlockId, TransactionId}` now uses unverified DTO conversions;
 
 ### Fixed